refactor: separate data card component begins from confirmed cases card
