+++ pages/cards/_card.vue
@@ -1,0 +1,17 @@
+<template>
+  <confirmed-cases-card />
+</template>
+
+<script>
+import ConfirmedCasesCard from '@/components/cards/ConfirmedCasesCard.vue'
+
+export default {
+  components: {
+    ConfirmedCasesCard
+  },
+  // layout: 'card'
+}
+</script>
+
+<style lang="scss">
+</style>
