+++ pages/index.vue
@@ -13,18 +13,7 @@
       :btn-text="'相談の手順を見る'"
     />
     <v-row class="DataBlock">
-      <v-col cols="12" md="6" class="DataCard">
-        <svg-card
-          title="検査陽性者の状況"
-          :title-id="'details-of-confirmed-cases'"
-          :date="headerItem.date"
-        >
-          <confirmed-cases-table
-            aria-label="検査陽性者の状況"
-            v-bind="confirmedCases"
-          />
-        </svg-card>
-      </v-col>
+      <confirmed-cases-card />
       <v-col cols="12" md="6" class="DataCard">
         <time-bar-chart
           title="陽性患者数"
@@ -106,10 +95,8 @@ import MetroData from '@/data/metro.json'
 import DataTable from '@/components/DataTable.vue'
 import formatGraph from '@/utils/formatGraph'
 import formatTable from '@/utils/formatTable'
-import formatConfirmedCases from '@/utils/formatConfirmedCases'
 import News from '@/data/news.json'
-import SvgCard from '@/components/SvgCard.vue'
-import ConfirmedCasesTable from '@/components/ConfirmedCasesTable.vue'
+import ConfirmedCasesCard from '@/components/cards/ConfirmedCasesCard.vue'
 
 export default {
   components: {
@@ -120,8 +107,7 @@ export default {
     WhatsNew,
     StaticInfo,
     DataTable,
-    SvgCard,
-    ConfirmedCasesTable
+    ConfirmedCasesCard,
   },
   data() {
     // 感染者数グラフ
@@ -153,8 +139,6 @@ export default {
     // const fatalitiesTable = formatTable(
     //   Data.patients.data.filter(patient => patient['備考'] === '死亡')
     // )
-    // 検査陽性者の状況
-    const confirmedCases = formatConfirmedCases(Data.main_summary)
 
     const sumInfoOfPatients = {
       lText: patientsGraph[
@@ -176,7 +160,6 @@ export default {
       inspectionsGraph,
       inspectionsItems,
       inspectionsLabels,
-      confirmedCases,
       sumInfoOfPatients,
       headerItem: {
         icon: 'mdi-chart-timeline-variant',
